refactor(anime): Split mappers, filters, and relation types
