+++ src/domains/anime/mappers/index.ts
@@ -10,6 +10,7 @@ export * from './anime-character-mapper'
 export * from './anime-external-mapper'
 export * from './anime-filters-mapper'
 export * from './anime-full-mapper'
+export * from './anime-full-mapper-helpers'
 export * from './anime-mapper'
 export * from './anime-music-mapper'
 export * from './anime-staff-mapper'
